backend/fastApi/requirements.txt
fastapi
snowflake-connector-python
boto3
python-dotenv
uvicorn
pandas